pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loggrep"
version = "1.0.0"
description = "A powerful command-line tool for searching log files with timestamp-aware filtering"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Sean", email = "your.email@example.com"}
]
maintainers = [
    {name = "Sean", email = "your.email@example.com"}
]
keywords = ["log", "grep", "timestamp", "search", "filter", "syslog", "logcat"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: System :: Systems Administration",
    "Topic :: Text Processing :: Filters",
    "Topic :: Utilities",
]
requires-python = ">=3.7"
dependencies = [
    "python-dateutil>=2.8.0",
    "colorama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/seanpor/Loggrep"
Repository = "https://github.com/seanpor/Loggrep"
Documentation = "https://github.com/seanpor/Loggrep#readme"
"Bug Tracker" = "https://github.com/seanpor/Loggrep/issues"
Changelog = "https://github.com/seanpor/Loggrep/releases"

[project.scripts]
loggrep = "loggrep.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
loggrep = ["py.typed"]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]